Switch from HMM to Lix
This should help make it easier to prevent this code from rotting as library versions change.

The versions I've committed here are sometimes forks / upgrades / different to what I had in hmm, so it might break things. I'm about to commit changes to source to update them.

Note: I'm using a local fork of ufront-orm and have not pushed the changes to Github yet.