Refactor so new Content rows are only created when "save" is clicked
Now we have routes so ContentEditorPage can work based on simply a templateID,
and will redirect you to the one with the GUID the first time it's saving.

We've also set up the routes for empty embed iframes and JSON content.

Now the whole existing feature set is working with tink_sql. Woo!