Add a package.json for the Dockerfile so it loads NPM modules correctly.
It might be worth having this copied from the parent directory, or generated by a macro.

For now I'm doing this as it's easier than figuring out how to get Lix to run in the docker container, which would be required if I just use the package.json from the top folder.